requirements/base.txt
Django
psycopg2
djangorestframework
django-redis
django-fcm

django-excel
pyexcel-xls
pyexcel-xlsx
pyexcel-ods